+++ Network/API/Telegram/Bot/Object/Update/Message/Content/File/Voice.hs
@@ -7,7 +7,6 @@ import "base" Data.Maybe (Maybe)
 import "base" Data.Function (($))
 import "base" Data.Functor ((<$>))
 import "base" Text.Show (Show)
-import "text" Data.Text (Text)
 
 import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Duration (Duration)
 import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.MIME (MIME)