+++ CHANGELOG.md
@@ -144,3 +144,4 @@
 * Define `Accessible` instance for newtypes to not use generated accessors records
 * Move `Caption`, `Duration` and `URI` modules into `Special` submodule
 * Define `MIME` newtype in `Special` submodule to replace vague `Text`
+* Rename `Size` to `Photosize` and move its content to `Photo` module